requirements.txt
requests
moviepy
Pillow
opencv-python
PyYAML
argparse
spacy